backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
pypdf2>=3.0.1
langchain>=0.0.335
langchain-community>=0.0.13
faiss-cpu>=1.12.0
sentence-transformers>=2.2.2
python-multipart>=0.0.6
beautifulsoup4>=4.12.2
lxml>=4.9.3
xmltodict>=0.13.0
openai>=1.3.0
ollama>=0.5.0
PyMuPDF>=1.20.0
Pillow>=9.0.0
python-dotenv>=1.0.0
defusedxml>=0.7.0
slowapi>=0.1.9
chromadb>=0.4.22
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.3
deep-translator>=1.11.4
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0 